service/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pytest
httpx
cyclonedx-bom
setuptools>=78.1.1